Remove mulligan
"Mulligan" as a specific part of the replay is an error : actions happens before, after mulligan, and cards are changed during. But in the end, it's still part of the "steps" of a round, making it something different overcomplicates the system (and make it rigid if mulligan phase changes in the future of Gwent).
It's changed for easier "step" element with a "mulligan" tag. "Mulligan" card changes for players are nothing more than "choice > move deck > move hand" phases.